06 blazor get socks

+++ Code/Domain/Models/Product.cs
@@ -3,7 +3,7 @@ namespace Shared.Models;
 public class Product
 {
     private ProductCard ProductCard { get; set; }
-    private long Id { get; set; }
+    private int Id { get; set; }
     private string Size { get; set; }
     private string Color { get; set; }
